Using typedi
