+++ package.json
@@ -2,7 +2,7 @@
   "name": "loset",
   "productName": "Loset",
   "version": "0.2.1",
-  "homepage": "https://github.com/yangukmo/loset",
+  "homepage": "https://loset.app",
   "private": true,
   "author": "yangukmo",
   "description": "Loset",
@@ -46,11 +46,11 @@
     "xterm-addon-search": "^0.4.0"
   },
   "devDependencies": {
-    "@types/node": "^13.7.1",
+    "@types/node": "12.12.27",
     "@types/uuid": "^3.4.7",
     "@types/vue-color": "^2.4.2",
-    "@typescript-eslint/eslint-plugin": "^2.19.2",
-    "@typescript-eslint/parser": "^2.19.2",
+    "@typescript-eslint/eslint-plugin": "^2.20.0",
+    "@typescript-eslint/parser": "^2.20.0",
     "@vue/cli-plugin-babel": "^4.2.2",
     "@vue/cli-plugin-eslint": "^4.2.2",
     "@vue/cli-plugin-typescript": "^4.2.2",
@@ -60,7 +60,7 @@
     "devtron": "^1.4.0",
     "electron": "^8.0.1",
     "eslint": "^6.8.0",
-    "eslint-plugin-vue": "^6.1.2",
+    "eslint-plugin-vue": "^6.2.1",
     "node-sass": "^4.13.1",
     "sass-loader": "^8.0.2",
     "typescript": "^3.7.5",
