+++ src/api/config/config-manager.ts
@@ -1,3 +1,7 @@
+import 'reflect-metadata'
+import { Service } from 'typedi'
+
+@Service()
 export default class ConfigManager {
 
 }
