+++ src/api/reflect-metadata.ts
@@ -1,0 +1,4 @@
+import 'reflect-metadata'
+
+export default class ReflectMetadata {
+}
